requirements.txt
ReactivePy==2.0.0.dev0
discord.py==1.5.1
varname==0.5.3
urllib3==1.26.2
python-dotenv==0.15.0
praw==7.1.0
prawcore==1.5.0
async-timeout==3.0.1
certifi==2020.11.8
attrs==20.3.0
chardet==3.0.4
multidict==5.0.2
pip==20.2.4
requests==2.25.0
setuptools==50.3.2
six==1.15.0
typing-extensions==3.7.4.3
update-checker==0.18.0
websocket-client==0.57.0
yarl==1.6.3
idna==2.10
executing==0.5.3
aiohttp==3.7.3
Pillow==8.0.1
python-dateutil==2.8.1